fixed wofi for scaling, endeavouros setup script

+++ scripts/arch_setup.sh
@@ -17,36 +17,35 @@ sudo pacman -S --needed --noconfirm \
 	openssh curl wget ripgrep fd lazygit \
 	nnn atool libarchive zip unzip trash-cli sshfs rclone fuse2 \
 	iwd wireless_tools wpa_supplicant \
-	neovim neovim-plugins python-pynvim
-
-kitty kitty-terminfo tree yay apparmor \
-	mtr
+	neovim neovim-plugins python-pynvim \
+	kitty kitty-terminfo tree yay apparmor \
+	mtr bluez-utils acpi network-manager-applet \
+	btop
 
 # hyprland and wayland apps
 sudo pacman -S --needed --noconfirm \
-	mako wofi \
+	mako wofi polkit-gnome \
 	hyprland xdg-desktop-portal-hyprland \
 	xdg-utils xdg-desktop-portal-gtk \
 	polkit-kde-agent polkit-qt6 gnome-keyring \
 	swaybg swayidle \
 	wl-clipboard wayland-utils \
-	qt5-wayland qt6-wayland \
+	qt5-wayland qt6-wayland qt5ct qt6ct \
 	pipewire wireplumber \
 	grim slurp
 
-yay -Syua --noconfirm --sudoloop
-yay -S --needed --noconfirm --sudoloop \
-	waybar-hyprland swaylock-effects
+yay -Syua --needed --noconfirm --sudoloop \
+	waybar-hyprland swaylock-effects nwg-look \
+	catppuccin-gtk-theme-mocha
+#catppuccin-gtk-theme-macchiato catppuccin-gtk-theme-frappe catppuccin-gtk-theme-latte
 
 # font and LaTeX installation
 sudo pacman -S --needed --noconfirm \
 	tectonic \
-	ttf-dejavu-nerd \
-	ttf-firacode-nerd \
-	ttf-font-awesome \
-	ttf-inconsolata-nerd \
+	ttf-dejavu-nerd ttf-inconsolata-nerd ttf-firacode-nerd \
 	ttf-nerd-fonts-symbols ttf-nerd-fonts-symbols-mono \
-	ttf-ubuntu-nerd ttf-ubuntu-mono-nerd
+	ttf-font-awesome \
+	awesome-terminal-fonts ttf-ubuntu-nerd ttf-ubuntu-mono-nerd
 
 # development stuff
 sudo pacman -S --needed --noconfirm \
@@ -56,10 +55,12 @@ sudo pacman -S --needed --noconfirm \
 	python-pip python-black python-poetry
 
 yay -S --needed --noconfirm --sudoloop \
-	podman-desktop \
 	opentofu
+#podman-desktop \
 
-sudo ln -s /usr/bin/tofu /usr/bin/terraform
+if [ ! -e "/usr/bin/terraform" ]; then
+	sudo ln -s /usr/bin/tofu /usr/bin/terraform
+fi
 
 # security stuff
 yay -S --needed --noconfirm --sudoloop \
